Implement proper max zoom
Zoom sensitivity is reduced at close zoom

Renderer fixes
- Fix preparing more than chunks specified
- Fix incorrect tiles_x to prepare calculation
- Resolved conversion warnings in renderer.cpp

+++ include/render/renderer_layer.h
@@ -8,6 +8,7 @@
 
 #include "jactorio.h"
 
+#include "core/convert.h"
 #include "core/coordinate_tuple.h"
 #include "core/data_type.h"
 #include "render/opengl/vertex_array.h"
@@ -226,7 +227,7 @@ namespace jactorio::render
     }
 
     inline uint32_t RendererLayer::Size() const noexcept {
-        return (writePtr_ - baseBuffer_) / kBaseValsPerElement;
+        return SafeCast<uint32_t>((writePtr_ - baseBuffer_) / kBaseValsPerElement);
     }
 } // namespace jactorio::render
